+++ src/Common/Mailer.php
@@ -94,13 +94,13 @@ class Mailer
 
         // validate
         if (!\SpoonFilter::isEmail($email['to_email'])) {
-            throw new Exception('Invalid e-mail address for recipient.');
+            throw new \Exception('Invalid e-mail address for recipient.');
         }
         if (!\SpoonFilter::isEmail($email['from_email'])) {
-            throw new Exception('Invalid e-mail address for sender.');
+            throw new \Exception('Invalid e-mail address for sender.');
         }
         if (!\SpoonFilter::isEmail($email['reply_to_email'])) {
-            throw new Exception('Invalid e-mail address for reply-to address.');
+            throw new \Exception('Invalid e-mail address for reply-to address.');
         }
 
         // build array